pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "gan-vs-sagan"
version = "1.0.0"
description = "Fair comparison of GAN vs Self-Attention GAN on CelebA-HQ"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy>=1.24",
    "Pillow>=9.0",
    "PyYAML>=6.0",
    "tqdm>=4.65",
    "tensorboard>=2.13",
    "clean-fid>=0.1.35",
    "scipy>=1.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["src*"]
